+++ CHANGELOG.md
@@ -10,12 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Changed minimum required Elixir version to 1.10
+- `CodestatsApi` namespace renamed to `CodestatsAPI`
 
 ## [0.1.0] - 2023-11-15
 
 ### Added
 
-- Added `CodestatsApi.get_user_data/1` interface
+- Added `CodestatsAPI.get_user_data/1` interface
 
 [unreleased]: https://github.com/general-CbIC/codestats_api/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/general-CbIC/codestats_api/releases/tag/v0.1.0